make scl as a separate deploy
