+++ package-engine.json
@@ -1,6 +1,6 @@
 {
     "name": "@kedoska/engine-slot",
-    "version": "0.0.3",
+    "version": "0.0.4",
     "description": "A library to build slot machines",
     "main": "./index.js",
     "types": "./index.d.ts",
