+++ package-scl.json
@@ -1,6 +1,6 @@
 {
     "name": "@kedoska/engine-slot-scl",
-    "version": "0.0.3",
+    "version": "0.0.4",
     "description": "A library of configurations for the engine-slot",
     "main": "./index.js",
     "types": "./index.d.ts",
