+++ publish.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-
+ORIGIN=$(pwd)
 npm run format
 
 if output=$(git status --porcelain) && [ "$output" ]; then
@@ -19,10 +19,13 @@ cd $ENGINE
 echo entering `pwd`
 npm publish
 
+cd $ORIGIN
+
 SCL=./dist/scl
-cp ./README.md $SCL
-cp ./package-engine.json $SCL/package.json
+cp ./README-SCL.md $SCL/README.md
+cp ./package-scl.json $SCL/package.json
 cd $SCL
+find -type f -name '*.spec.*' -delete
 echo entering `pwd`
 npm publish
